requirements.txt
click8==8.0.1
DateTime==4.3
names==0.3.0
pytz==2020.1
tabulate==0.8.7
zope.interface==5.1.0
